Add 'packed index' API, includes opaque bit if needed
The 'block index' is still a small integer used as an array index into
blockProps, but the new 'packed block index' may include packed
bitflags. Specifically, if voxel-mesher is loaded, then (by default) the
1<<15 bit will be set from getPackedBlockIndex(), iff the 'transparent'
property is not set. This is required for correct voxel meshing.
getBlockName() now also masks out this bit if it is present.